Code structure refactoring * Split run.py to separated modules based on their functionality; * Move logger to a module that doesn't conflict with the standard module by name; * Attempts to enter correct password and exceptions intercept in main
