+++ src/w3plex/__init__.py
@@ -1,5 +1,5 @@
 from lazyplex import *  # noqa: F403
 
 from .core import * # noqa: F403
-from .logging import * # noqa: F403
+from .log import * # noqa: F403
 from .exceptions import * # noqa: F403
