+++ src/w3plex/constants.py
@@ -2,3 +2,4 @@ CONTEXT_CHAINS_KEY = '_chains'
 CONTEXT_CONFIG_KEY = '_config'
 CONTEXT_EXTRAS_KEY = '_extras'
 CONTEXT_LOGGER_KEY = '_logger'
+APPLICATIONS_CFG_KEY = 'applications'
